wrap category menu into a component
